remove @getInvertedPairedCharacters()
Caching @invertedPairedCharacters() is an idiotic attempt at optimizing
an operation (inverting a 6-10 key hash-map) that isn't even
computationally intensive. Plus, when @toggleQuotes() is called, it
mutates @pairedCharacters, making @invertedPairedCharacters incorrect.

You could try to handle cache invalidation when @toggleQuotes() is
called, but the operation of checking to see if the inverted map should
be updated or not would probably take more time than the inversion
itself. Thus, it's shorter & cheaper to use _.values() & skip inversion
altogether.